moves verification and execution builders into separate files
